+++ android/src/main/kotlin/ch/waio/pro_video_editor/src/features/render/helpers/CompositionBuilder.kt
@@ -94,6 +94,7 @@ class CompositionBuilder(
                     height = imageLayer.height,
                     rotation = imageLayer.rotation,
                     loop = imageLayer.loop,
+                    animationOffsetUs = imageLayer.animationOffsetUs,
                     animations = imageLayer.animations
                 )
             })